// types/batch-table.ts
export type BatchTableToolConfigurationId = Tagged<
	number,
	"BatchTableToolConfigurationId"
>;
export type BatchTableToolSettingsId = Tagged<
	number,
	"BatchTableToolSettingsId"
>;
export type BatchTableColumnIndex = Tagged<number, "BatchTableColumnIndex">;
export type BatchTableColumnUuid = Tagged<string, "BatchTableColumnUuid">;
export type BatchTableCellUuid = Tagged<string, "BatchTableCellUuid">;
export type BatchTableColumnId = Tagged<number, "BatchTableColumnId">;
export type BatchTableRowIndex = Tagged<number, "BatchTableRowIndex">;
export type BatchTableRowUuid = Tagged<string, "BatchTableRowUuid">;
export type BatchTableCellId = Tagged<number, "BatchTableCellId">;
export type BatchTableRowId = Tagged<number, "BatchTableRowId">;
export type BatchTableToolId = Tagged<number, "BatchTableToolId">;

type SingleLineTextInfo = Readonly<EmptyObject>;

type FileColumnType = Readonly<EmptyObject>;

type LongTextInfo = Readonly<EmptyObject>;

type NumberInfo = Readonly<{
	is_integer: boolean;
}>;

type BatchTableMetadataColumnTypeSpecificInfo =
	| SingleLineTextInfo
	| FileColumnType
	| LongTextInfo
	| NumberInfo;

export enum DerivationType {
	JSON_COLUMN_DATA = "JSON_COLUMN_DATA",
	TOOL_OUTPUTS = "TOOL_OUTPUTS",
}

export type BatchTableColumnChildFilter = {
	value_operator: ColumnOptionsValueOperator;
	column_id: BatchTableColumnId | undefined;
	caseSensitive: boolean;
	value: unknown;
};

export enum BatchTableMetadataColumnType {
	SINGLE_LINE_TEXT = "SINGLE_LINE_TEXT",
	LONG_TEXT = "LONG_TEXT",
	BOOLEAN = "BOOLEAN",
	NUMBER = "NUMBER",
	FILE = "FILE",
	JSON = "JSON",
}

export const BATCH_TABLE_METADATA_COLUMN_TYPE_OPTIONS = Object.values(
	BatchTableMetadataColumnType,
);

export type BatchTableConditionalExecutionFilterGroup = {
	children: Array<
		BatchTableColumnChildFilter | BatchTableConditionalExecutionFilterGroup
	>;
	filter_operator: FilterOperator;
};

export type BatchTableColumn = Nullable<{
	derived_from_column: {
		uuid: BatchTableColumnUuid;
		id: BatchTableColumnId;
	};
	column_type_specific_info: BatchTableMetadataColumnTypeSpecificInfo;
	execution_condition: BatchTableConditionalExecutionFilterGroup;
	column_type: BatchTableMetadataColumnType;
	tool_settings: BatchTableToolSettings;
	last_modified_by: BetterbrainUser;
	derivation_type: DerivationType;
	derivation_path: Array<string>;
	created_by: BetterbrainUser;
	updated_at: ISODateString;
	created_at: ISODateString;
	default_value: unknown;
	is_derived: boolean;
	description: string;
	use_ai: boolean;
	prompt: string;
	name: string;
}> & {
	column_format: {
		isVisible: boolean;
		width: number;
	};
	column_index: BatchTableColumnIndex;
	uuid: BatchTableColumnUuid;
	id: BatchTableColumnId;
};

export enum BatchTableCellFormatPattern {
	DateAndTime = "Date and Time",
	Financial = "Financial",
	Duration = "Duration",
	Percent = "Percent",
	Number = "Number",
	Date = "Date",
	Time = "Time",
	Text = "Text",
	None = "None",
}
// ...
export type ToolConfiguration = Nullable<{
	/** Keys are the input names */
	inputs: Record<string, unknown>;
	last_modified_at: ISODateString;
	created_at: ISODateString;
	updated_at: ISODateString;
}> & {
	id: BatchTableToolConfigurationId;
	tool_priority: number;
	tool: Tool;
};

enum SoftwareTools {
	GoogleDrive = "GOOGLE_DRIVE",
}

export type GeneralToolInput = {
	type: "boolean" | "array<string>" | "string" | "integer";
	can_be_inferred: boolean;
	/** String or array or integer or number or boolean or object: */
	default_value: unknown;
	is_required: boolean;
	has_default: boolean;
	description: string;
	name: string;
};

export type Tool = {
	depends_on_software_tools: Array<SoftwareTools>;
	// can_be_used_with_entity_column: boolean; // Entity column is not used anymore
	is_available_in_batch_table: boolean;
	metadata: Record<string, unknown>;
	outputs: Record<string, unknown>;
	inputs: Array<GeneralToolInput>;
	last_modified_at: ISODateString;
	updated_at: ISODateString;
	created_at: ISODateString;
	description: string;
	user_name: string;
	name: string;
	id: BatchTableToolId;
};

export enum BatchTableToolSettingsInheritanceType {
	INHERIT = "INHERIT",
	CUSTOM = "CUSTOM",
}

export type BatchTableToolSettings = Nullable<{
	inheritance_type: BatchTableToolSettingsInheritanceType;
	tool_configurations: Array<ToolConfiguration>;
	source_columns: Array<BatchTableColumnId>;
	use_all_columns: boolean;
}> & {
	id: BatchTableToolSettingsId;
};